requirements.txt
python-telegram-bot==20.07
requests
flask